working version for vite
