--- src/App.js
@@ -1,16 +1,0 @@
-import LayoutElements from './layoutElemets';
-import AuthElements from './firebaseAuth';
-
-
-import './App.css';
-import './exp.css';
-
-function App() {
-  return (
-    <div className="App">
-      <AuthElements.AppLogin />
-    </div>
-  );
-}
-
-export default App;
