+++ src/firebaseAuth.css
@@ -3,7 +3,7 @@
     --header-height: 56px;
     --footer-height: 80px; 
     --line-height: 2px;
-    --gaps: 4px; /* line 총합 */
+    --gaps: 4px; /* Total line spacing */
     --main-color: #96E36C;
     --light-color: #bcbcbcf3;
     --background-default: #3a3c32;
