+++ src/iconElements.jsx
@@ -1,0 +1,121 @@
+import './iconElements.css'
+
+function HomeIcon() {
+    return (
+        <svg className="FooterIcon HomeIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M1.58594 10.5453L10.9393 1.43968C11.0786 1.30029 11.244 1.18971 11.426 1.11427C11.6081 1.03883 11.8032 1 12.0003 1C12.1973 1 12.3925 1.03883 12.5745 1.11427C12.7566 1.18971 12.922 1.30029 13.0613 1.43968L22.4146 10.5453" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M16.5 4.62866V3.12866H20V7.84375" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M3 9.5V18.069V21.6552C3 22.0118 3.1724 22.3539 3.47928 22.6061C3.78616 22.8583 4.20237 23 4.63636 23H19.3636C19.7976 23 20.2138 22.8583 20.5207 22.6061C20.8276 22.3539 21 22.0118 21 21.6552V9.5" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M15 23V18C15 16.3431 13.6569 15 12 15V15C10.3431 15 9 16.3431 9 18V23" stroke="currentColor" strokeWidth="1.5"/>
+            <path d="M9 10.25C8.58579 10.25 8.25 10.5858 8.25 11C8.25 11.4142 8.58579 11.75 9 11.75V10.25ZM15 11.75C15.4142 11.75 15.75 11.4142 15.75 11C15.75 10.5858 15.4142 10.25 15 10.25V11.75ZM9 11.75H15V10.25H9V11.75Z" fill="currentColor"/>
+            <path d="M9 7.25C8.58579 7.25 8.25 7.58579 8.25 8C8.25 8.41421 8.58579 8.75 9 8.75V7.25ZM15 8.75C15.4142 8.75 15.75 8.41421 15.75 8C15.75 7.58579 15.4142 7.25 15 7.25V8.75ZM9 8.75H15V7.25H9V8.75Z" fill="currentColor"/>
+        </svg>
+    );
+}
+
+function RankingIcon() {
+    return (
+        <svg className="FooterIcon RankingIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M22 7L14.1502 14.939C14.0125 15.0784 13.8489 15.189 13.6689 15.2644C13.4889 15.3398 13.296 15.3787 13.1011 15.3787C12.9063 15.3787 12.7133 15.3398 12.5333 15.2644C12.3533 15.189 12.1898 15.0784 12.052 14.939L8.95919 11.811C8.82146 11.6716 8.65791 11.561 8.47791 11.4856C8.2979 11.4102 8.10496 11.3713 7.91011 11.3713C7.71526 11.3713 7.52232 11.4102 7.34232 11.4856C7.16231 11.561 6.99877 11.6716 6.86103 11.811L2 16.1147" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M2 1.5V22.5" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M22 12V7H17" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+function ChallengeIcon() {
+    return (
+        <svg className="FooterIcon ChallengeIcon" viewBox="-0.5 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M12 23.25C18.2132 23.25 23.25 18.2132 23.25 12C23.25 5.7868 18.2132 0.75 12 0.75C5.7868 0.75 0.75 5.7868 0.75 12C0.75 18.2132 5.7868 23.25 12 23.25Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M13.3337 6.00201L8.07871 12.1202C8.03772 12.1688 8.01144 12.228 8.00298 12.2909C7.99451 12.3539 8.0042 12.4179 8.03091 12.4755C8.05763 12.5332 8.10025 12.5819 8.15376 12.6162C8.20727 12.6504 8.26945 12.6686 8.33297 12.6687H10.6666V18.002L15.9216 11.8847C15.9624 11.8361 15.9886 11.7768 15.9971 11.7139C16.0055 11.6509 15.9958 11.5868 15.9691 11.5292C15.9424 11.4716 15.8999 11.4227 15.8464 11.3884C15.793 11.354 15.7308 11.3356 15.6673 11.3353H13.3337V6.00201Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+function GroupIcon() {
+    return (
+        <svg className="FooterIcon GroupIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M12 6C13.1046 6 14 5.10457 14 4C14 2.89543 13.1046 2 12 2C10.8954 2 10 2.89543 10 4C10 5.10457 10.8954 6 12 6Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M15.75 14V12C15.75 11.0054 15.3549 10.0516 14.6517 9.34835C13.9484 8.64509 12.9946 8.25 12 8.25C11.0054 8.25 10.0516 8.64509 9.34835 9.34835C8.64509 10.0516 8.25 11.0054 8.25 12V14H9.75L10.5 20H13.5L14.25 14H15.75Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M5 8C6.10457 8 7 7.10457 7 6C7 4.89543 6.10457 4 5 4C3.89543 4 3 4.89543 3 6C3 7.10457 3.89543 8 5 8Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M3.94598 9.79254C3.35733 9.88089 2.79813 10.108 2.31457 10.4551C1.831 10.8022 1.43693 11.2593 1.16488 11.7887C0.892841 12.3182 0.750633 12.9047 0.75 13.5V14H2.25L3 19H6L6.376 16.494" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M19 8C20.1046 8 21 7.10457 21 6C21 4.89543 20.1046 4 19 4C17.8954 4 17 4.89543 17 6C17 7.10457 17.8954 8 19 8Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M20.054 9.79254C20.6427 9.88089 21.2019 10.108 21.6855 10.4551C22.169 10.8022 22.5631 11.2593 22.8351 11.7887C23.1072 12.3182 23.2494 12.9047 23.25 13.5V14H21.75L21 19H18L17.624 16.494" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M4 23H20" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+function UserIcon() {
+    return (
+        <svg className="FooterIcon UserIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M2.01557 21.4673L2.75667 21.5824L2.01557 21.4673ZM21.9844 21.4673L21.2433 21.5824L21.9844 21.4673ZM2.75667 21.5824C3.44504 17.1525 7.31691 13.75 12 13.75V12.25C6.57737 12.25 2.07637 16.1915 1.27446 21.3521L2.75667 21.5824ZM12 13.75C16.6831 13.75 20.555 17.1525 21.2433 21.5824L22.7255 21.3521C21.9236 16.1915 17.4226 12.25 12 12.25V13.75ZM20.5481 22.25H3.45185V23.75H20.5481V22.25ZM1.27446 21.3521C1.05985 22.7332 2.22557 23.75 3.45185 23.75V22.25C2.96139 22.25 2.71029 21.8809 2.75667 21.5824L1.27446 21.3521ZM21.2433 21.5824C21.2897 21.8809 21.0386 22.25 20.5481 22.25V23.75C21.7744 23.75 22.9402 22.7332 22.7255 21.3521L21.2433 21.5824Z" fill="currentColor"/>
+            <path d="M16.25 6C16.25 8.34721 14.3472 10.25 12 10.25V11.75C15.1756 11.75 17.75 9.17564 17.75 6H16.25ZM12 10.25C9.65279 10.25 7.75 8.34721 7.75 6H6.25C6.25 9.17564 8.82436 11.75 12 11.75V10.25ZM7.75 6C7.75 3.65279 9.65279 1.75 12 1.75V0.25C8.82436 0.25 6.25 2.82436 6.25 6H7.75ZM12 1.75C14.3472 1.75 16.25 3.65279 16.25 6H17.75C17.75 2.82436 15.1756 0.25 12 0.25V1.75Z" fill="currentColor"/>
+            <path d="M6.90014 4.80875C7.75812 5.46874 9.25809 6.27245 11.0481 6.61187C12.8523 6.954 14.9861 6.8289 17.0362 5.56732L16.2501 4.28983C14.5859 5.31396 12.8506 5.42696 11.3275 5.13813C9.79009 4.8466 8.50434 4.15031 7.8147 3.61982L6.90014 4.80875Z" fill="currentColor"/>
+        </svg>
+    );
+}
+
+function NotificationIcon() {
+    return (
+        <svg className="HeaderIcon NotificationIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M10.0098 21.7821C10.2782 22.6436 11.0982 23.2492 12.0025 23.2492C12.9069 23.2492 13.7269 22.6436 13.9953 21.7821" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M12 3V0.75" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M12 3C16.1682 3 19.5 6.49157 19.5 10.619C19.5 17.7769 21 19 21 19H3C3 19 4.5 17.0536 4.5 10.619C4.5 6.51979 7.86509 3 12 3Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+function SettingsIcon() {
+    return (
+        <svg className="HeaderIcon SettingsIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M10.5459 2.438C10.7294 2.6417 10.9537 2.80457 11.2042 2.91605C11.4546 3.02754 11.7258 3.08514 11.9999 3.08514C12.2741 3.08514 12.5452 3.02754 12.7957 2.91605C13.0462 2.80457 13.2704 2.6417 13.4539 2.438L14.3999 1.4C14.671 1.09995 15.0284 0.891146 15.4229 0.802265C15.8174 0.713384 16.2298 0.748776 16.6034 0.903585C16.977 1.05839 17.2935 1.32503 17.5095 1.6669C17.7256 2.00877 17.8305 2.40912 17.8099 2.813L17.7389 4.213C17.7252 4.48611 17.7688 4.75907 17.867 5.01429C17.9652 5.26951 18.1157 5.50134 18.3089 5.69484C18.5021 5.88834 18.7337 6.03922 18.9888 6.13777C19.2439 6.23631 19.5168 6.28034 19.7899 6.267L21.1899 6.196C21.5935 6.17611 21.9934 6.28152 22.3348 6.49779C22.6762 6.71407 22.9423 7.03062 23.0967 7.40407C23.2511 7.77752 23.2863 8.18958 23.1973 8.58379C23.1084 8.97799 22.8997 9.33504 22.5999 9.606L21.5579 10.546C21.3545 10.7298 21.1918 10.9542 21.0805 11.2047C20.9692 11.4552 20.9117 11.7264 20.9117 12.0005C20.9117 12.2747 20.9692 12.5458 21.0805 12.7963C21.1918 13.0468 21.3545 13.2712 21.5579 13.455L22.5999 14.395C22.9 14.6661 23.1088 15.0235 23.1977 15.418C23.2865 15.8125 23.2511 16.2249 23.0963 16.5985C22.9415 16.9721 22.6749 17.2886 22.333 17.5046C21.9911 17.7206 21.5908 17.8256 21.1869 17.805L19.7869 17.734C19.5132 17.7199 19.2396 17.7635 18.9838 17.8618C18.7279 17.9601 18.4956 18.1111 18.3018 18.3049C18.108 18.4987 17.9571 18.731 17.8587 18.9869C17.7604 19.2427 17.7168 19.5163 17.7309 19.79L17.8019 21.19C17.8198 21.5919 17.7136 21.9895 17.4976 22.3289C17.2817 22.6682 16.9665 22.9329 16.5949 23.0869C16.2233 23.2409 15.8133 23.2768 15.4206 23.1897C15.0279 23.1026 14.6715 22.8967 14.3999 22.6L13.4589 21.559C13.2752 21.3556 13.0509 21.193 12.8005 21.0817C12.55 20.9704 12.279 20.9128 12.0049 20.9128C11.7308 20.9128 11.4598 20.9704 11.2094 21.0817C10.9589 21.193 10.7346 21.3556 10.5509 21.559L9.60592 22.6C9.33473 22.8981 8.97821 23.1053 8.58497 23.1934C8.19172 23.2814 7.78088 23.2461 7.40846 23.0921C7.03604 22.9382 6.72016 22.6731 6.50389 22.3331C6.28761 21.993 6.18147 21.5946 6.19992 21.192L6.27192 19.792C6.286 19.5183 6.24246 19.2447 6.14411 18.9889C6.04577 18.733 5.89482 18.5007 5.70102 18.3069C5.50722 18.1131 5.27489 17.9622 5.01907 17.8638C4.76325 17.7655 4.48963 17.7219 4.21592 17.736L2.81592 17.807C2.4122 17.8281 2.01186 17.7237 1.66986 17.5081C1.32785 17.2926 1.06095 16.9764 0.905801 16.6031C0.750649 16.2298 0.714851 15.8176 0.803323 15.4232C0.891794 15.0287 1.1002 14.6713 1.39992 14.4L2.44092 13.46C2.64437 13.2762 2.807 13.0518 2.91831 12.8013C3.02962 12.5508 3.08713 12.2797 3.08713 12.0055C3.08713 11.7314 3.02962 11.4602 2.91831 11.2097C2.807 10.9592 2.64437 10.7348 2.44092 10.551L1.39992 9.606C1.10117 9.33501 0.893402 8.97836 0.80502 8.58481C0.716638 8.19126 0.751954 7.78002 0.906141 7.4073C1.06033 7.03458 1.32586 6.71858 1.66644 6.50248C2.00702 6.28639 2.40603 6.18075 2.80892 6.2L4.20892 6.271C4.48315 6.28542 4.75734 6.242 5.0137 6.14354C5.27005 6.04509 5.50282 5.89382 5.69691 5.69955C5.89099 5.50528 6.04204 5.27235 6.14024 5.0159C6.23844 4.75946 6.2816 4.48522 6.26692 4.211L6.19992 2.81C6.18102 2.40727 6.28685 2.00851 6.50296 1.66815C6.71907 1.32778 7.03494 1.06238 7.40746 0.908167C7.77998 0.753953 8.19101 0.718434 8.58446 0.806457C8.97791 0.89448 9.33462 1.10176 9.60592 1.4L10.5459 2.438Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M12 16C14.2091 16 16 14.2091 16 12C16 9.79086 14.2091 8 12 8C9.79086 8 8 9.79086 8 12C8 14.2091 9.79086 16 12 16Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+function FitnessIcon() {
+    return (
+        <svg className="FooterIcon FitnessIcon" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M16.9977 1.2416L16.4455 0.689198L16.4453 0.689365L16.9977 1.2416ZM16.9977 3.45054L16.4453 4.00277L16.4454 4.00286L16.9977 3.45054ZM21.5499 8.00277L22.1024 7.45058L22.1023 7.45045L21.5499 8.00277ZM23.7599 5.79383L24.3124 5.24164L24.3122 5.24145L23.7599 5.79383ZM19.2066 1.2416L19.7589 0.689218L19.7589 0.689198L19.2066 1.2416ZM18.1022 0.784211V0.00311472V0.784211ZM16.4453 0.689365C16.006 1.1288 15.7592 1.72471 15.7592 2.34607H17.3214C17.3214 2.13895 17.4037 1.94031 17.5501 1.79383L16.4453 0.689365ZM15.7592 2.34607C15.7592 2.96743 16.006 3.56334 16.4453 4.00277L17.5501 2.8983C17.4037 2.75183 17.3214 2.55319 17.3214 2.34607H15.7592ZM16.4454 4.00286L20.9976 8.55509L22.1023 7.45045L17.55 2.89822L16.4454 4.00286ZM20.9975 8.55496C21.215 8.77262 21.4733 8.94529 21.7576 9.06312L22.3558 7.61999C22.261 7.5807 22.1749 7.52314 22.1024 7.45058L20.9975 8.55496ZM21.7576 9.06312C22.0419 9.18096 22.3466 9.24164 22.6544 9.24172L22.6547 7.67953C22.5522 7.6795 22.4506 7.65927 22.3558 7.61999L21.7576 9.06312ZM22.6544 9.24172C22.9621 9.24179 23.2669 9.18125 23.5512 9.06355L22.9537 7.62013C22.8589 7.65937 22.7573 7.67955 22.6547 7.67953L22.6544 9.24172ZM23.5512 9.06355C23.8356 8.94585 24.0939 8.77329 24.3116 8.55574L23.2072 7.45084C23.1346 7.52337 23.0485 7.58089 22.9537 7.62013L23.5512 9.06355ZM24.3116 8.55574C24.5292 8.33818 24.7019 8.07989 24.8198 7.7956L23.3766 7.19743C23.3373 7.2922 23.2798 7.37831 23.2072 7.45084L24.3116 8.55574ZM24.8198 7.7956C24.9376 7.51131 24.9983 7.20659 24.9983 6.89885L23.4362 6.89848C23.4361 7.00107 23.4159 7.10265 23.3766 7.19743L24.8198 7.7956ZM24.9983 6.89885C24.9984 6.59111 24.9379 6.28636 24.8202 6.00202L23.3768 6.5995C23.416 6.6943 23.4362 6.79589 23.4362 6.89848L24.9983 6.89885ZM24.8202 6.00202C24.7025 5.71768 24.5299 5.4593 24.3124 5.24164L23.2075 6.34602C23.28 6.41857 23.3375 6.50471 23.3768 6.5995L24.8202 6.00202ZM24.3122 5.24145L19.7589 0.689218L18.6544 1.79398L23.2077 6.34621L24.3122 5.24145ZM19.7589 0.689198C19.3195 0.249899 18.7235 0.00311472 18.1022 0.00311472V1.56531C18.3093 1.56531 18.5079 1.64757 18.6544 1.794L19.7589 0.689198ZM18.1022 0.00311472C17.4808 0.00311472 16.8849 0.249899 16.4455 0.689198L17.5499 1.794C17.6964 1.64757 17.8951 1.56531 18.1022 1.56531V0.00311472Z" fill="currentColor"/>
+            <path d="M12.5781 1.24057L13.1303 1.79307L13.1314 1.79193L12.5781 1.24057ZM12.5781 3.45055L13.1304 2.89823L13.1303 2.89806L12.5781 3.45055ZM21.5503 12.4227L20.998 12.9751L20.9985 12.9756L21.5503 12.4227ZM23.7593 10.2096L23.2069 10.7619L23.2075 10.7625L23.7593 10.2096ZM14.7912 1.24057L14.2379 1.79193L14.2389 1.79286L14.7912 1.24057ZM13.6847 0.781096V0V0.781096ZM12.026 0.688076C11.8082 0.90569 11.6355 1.16408 11.5176 1.44848L12.9608 2.04653C13.0001 1.95173 13.0577 1.8656 13.1303 1.79307L12.026 0.688076ZM11.5176 1.44848C11.3998 1.73288 11.3391 2.03771 11.3391 2.34556H12.9013C12.9013 2.24294 12.9215 2.14133 12.9608 2.04653L11.5176 1.44848ZM11.3391 2.34556C11.3391 2.65341 11.3998 2.95825 11.5176 3.24265L12.9608 2.64459C12.9215 2.54979 12.9013 2.44818 12.9013 2.34556H11.3391ZM11.5176 3.24265C11.6355 3.52704 11.8082 3.78543 12.026 4.00305L13.1303 2.89806C13.0577 2.82552 13.0001 2.73939 12.9608 2.64459L11.5176 3.24265ZM12.0258 4.00287L20.998 12.9751L22.1026 11.8704L13.1304 2.89823L12.0258 4.00287ZM20.9985 12.9756C21.2164 13.1931 21.475 13.3655 21.7596 13.4831L22.3561 12.0393C22.2611 12 22.1748 11.9425 22.1021 11.8699L20.9985 12.9756ZM21.7596 13.4831C22.0442 13.6006 22.3491 13.661 22.657 13.6607L22.6555 12.0985C22.5528 12.0986 22.451 12.0785 22.3561 12.0393L21.7596 13.4831ZM22.657 13.6607C22.9649 13.6604 23.2697 13.5995 23.554 13.4814L22.9548 12.0387C22.86 12.0781 22.7583 12.0984 22.6555 12.0985L22.657 13.6607ZM23.554 13.4814C23.8384 13.3633 24.0967 13.1904 24.3142 12.9725L23.2085 11.8689C23.1359 11.9416 23.0497 11.9993 22.9548 12.0387L23.554 13.4814ZM24.3142 12.9725C24.5317 12.7545 24.7041 12.4959 24.8217 12.2114L23.3778 11.6149C23.3386 11.7098 23.2811 11.7961 23.2085 11.8689L24.3142 12.9725ZM24.8217 12.2114C24.9392 11.9268 24.9996 11.6219 24.9993 11.314L23.4371 11.3155C23.4372 11.4182 23.4171 11.5199 23.3778 11.6149L24.8217 12.2114ZM24.9993 11.314C24.999 11.0061 24.9381 10.7013 24.82 10.4169L23.3773 11.0161C23.4167 11.111 23.437 11.2127 23.4371 11.3155L24.9993 11.314ZM24.82 10.4169C24.7019 10.1326 24.529 9.8743 24.311 9.6568L23.2075 10.7625C23.2802 10.8351 23.3379 10.9212 23.3773 11.0161L24.82 10.4169ZM24.3116 9.65735L15.3436 0.688285L14.2389 1.79286L23.2069 10.7619L24.3116 9.65735ZM15.3445 0.689212C15.1268 0.470755 14.8681 0.297422 14.5833 0.179151L13.9842 1.62191C14.0792 1.66133 14.1654 1.71911 14.2379 1.79193L15.3445 0.689212ZM14.5833 0.179151C14.2985 0.0608808 13.9931 0 13.6847 0V1.56219C13.7875 1.56219 13.8893 1.58249 13.9842 1.62191L14.5833 0.179151ZM13.6847 0C13.3763 0 13.0709 0.0608808 12.7861 0.179151L13.3851 1.62191C13.4801 1.58249 13.5819 1.56219 13.6847 1.56219V0ZM12.7861 0.179151C12.5012 0.297422 12.2425 0.470755 12.0248 0.689212L13.1314 1.79193C13.204 1.71911 13.2902 1.66133 13.3851 1.62191L12.7861 0.179151Z" fill="currentColor"/>
+            <path d="M1.24154 16.9979L1.79394 17.5501L1.79403 17.55L1.24154 16.9979ZM3.45152 16.9979L2.89902 17.55L2.8992 17.5502L3.45152 16.9979ZM8.00375 21.5501L8.56566 21.0074L8.55606 20.9978L8.00375 21.5501ZM5.79481 23.759L5.24247 24.3115L5.25217 24.3209L5.79481 23.759ZM1.24154 19.2068L0.689134 19.7591L0.689281 19.7592L1.24154 19.2068ZM0.784147 18.1023H0.00305116H0.784147ZM1.79403 17.55C1.86657 17.4774 1.9527 17.4199 2.0475 17.3806L1.44944 15.9374C1.16504 16.0552 0.906653 16.228 0.689041 16.4457L1.79403 17.55ZM2.0475 17.3806C2.1423 17.3413 2.24391 17.3211 2.34653 17.3211V15.7589C2.03868 15.7589 1.73384 15.8195 1.44944 15.9374L2.0475 17.3806ZM2.34653 17.3211C2.44915 17.3211 2.55076 17.3413 2.64555 17.3806L3.24361 15.9374C2.95921 15.8195 2.65438 15.7589 2.34653 15.7589V17.3211ZM2.64555 17.3806C2.74035 17.4199 2.82648 17.4774 2.89902 17.55L4.00401 16.4457C3.7864 16.228 3.52801 16.0552 3.24361 15.9374L2.64555 17.3806ZM2.8992 17.5502L7.45143 22.1024L8.55606 20.9978L4.00384 16.4456L2.8992 17.5502ZM7.44191 22.0927C7.58419 22.2401 7.66292 22.4374 7.66114 22.6422L9.22328 22.6557C9.22862 22.0413 8.99243 21.4494 8.56558 21.0075L7.44191 22.0927ZM7.66114 22.6422C7.65937 22.847 7.57722 23.0429 7.4324 23.1877L8.53703 24.2923C8.9715 23.8579 9.21794 23.2701 9.22328 22.6557L7.66114 22.6422ZM7.4324 23.1877C7.28757 23.3325 7.09167 23.4147 6.88687 23.4164L6.90044 24.9786C7.51484 24.9732 8.10257 24.7268 8.53703 24.2923L7.4324 23.1877ZM6.88687 23.4164C6.68206 23.4182 6.48476 23.3395 6.33744 23.1972L5.25217 24.3209C5.69412 24.7477 6.28604 24.9839 6.90044 24.9786L6.88687 23.4164ZM6.34706 23.2067L1.79379 18.6544L0.689281 19.7592L5.24255 24.3114L6.34706 23.2067ZM1.79394 18.6546C1.6475 18.5081 1.56524 18.3095 1.56524 18.1023H0.00305116C0.00305116 18.7237 0.249836 19.3196 0.689134 19.7591L1.79394 18.6546ZM1.56524 18.1023C1.56524 17.8952 1.6475 17.6966 1.79394 17.5501L0.689134 16.4456C0.249835 16.8851 0.00305116 17.481 0.00305116 18.1023H1.56524Z" fill="currentColor"/>
+            <path d="M1.24109 12.5779L1.79219 13.1315L1.79333 13.1303L1.24109 12.5779ZM3.45003 12.5779L4.00235 12.0256L4.00227 12.0255L3.45003 12.5779ZM12.4222 21.5501L12.9749 20.9982L12.9745 20.9978L12.4222 21.5501ZM10.2091 23.7601L10.7618 23.2082L10.7614 23.2078L10.2091 23.7601ZM1.24109 14.7921L1.79341 14.2397L1.79219 14.2385L1.24109 14.7921ZM0.781096 13.685H1.56219H0.781096ZM1.79333 13.1303C1.93981 12.9839 2.13845 12.9016 2.34556 12.9016V11.3394C1.72421 11.3394 1.12829 11.5862 0.688859 12.0255L1.79333 13.1303ZM2.34556 12.9016C2.55268 12.9016 2.75132 12.9839 2.8978 13.1303L4.00227 12.0255C3.56284 11.5862 2.96692 11.3394 2.34556 11.3394V12.9016ZM2.89771 13.1302L11.8699 22.1024L12.9745 20.9978L4.00235 12.0256L2.89771 13.1302ZM11.8695 22.102C12.0162 22.2489 12.0985 22.4481 12.0984 22.6557L13.6606 22.6568C13.661 22.0349 13.4144 21.4382 12.9749 20.9982L11.8695 22.102ZM12.0984 22.6557C12.0982 22.8632 12.0156 23.0623 11.8687 23.2089L12.9726 24.3144C13.4127 23.8749 13.6601 23.2787 13.6606 22.6568L12.0984 22.6557ZM11.8687 23.2089C11.7218 23.3556 11.5227 23.4379 11.3151 23.4378L11.314 25C11.9359 25.0004 12.5325 24.7538 12.9726 24.3144L11.8687 23.2089ZM11.3151 23.4378C11.1075 23.4377 10.9085 23.3551 10.7618 23.2082L9.65641 24.312C10.0959 24.7521 10.6921 24.9996 11.314 25L11.3151 23.4378ZM10.7614 23.2078L1.79341 14.2397L0.688775 15.3444L9.6568 24.3124L10.7614 23.2078ZM1.79219 14.2385C1.71929 14.166 1.66145 14.0797 1.62198 13.9847L0.179364 14.5841C0.297773 14.8691 0.471305 15.1279 0.689997 15.3456L1.79219 14.2385ZM1.62198 13.9847C1.58251 13.8897 1.56219 13.7879 1.56219 13.685H0C0 13.9936 0.0609545 14.2991 0.179364 14.5841L1.62198 13.9847ZM1.56219 13.685C1.56219 13.5821 1.58251 13.4803 1.62198 13.3853L0.179364 12.7859C0.0609545 13.0708 0 13.3764 0 13.685H1.56219ZM1.62198 13.3853C1.66145 13.2903 1.71929 13.204 1.79219 13.1315L0.689997 12.0244C0.471305 12.2421 0.297773 12.5009 0.179364 12.7859L1.62198 13.3853Z" fill="currentColor"/>
+            <path d="M6.83154 15.9598L6.27922 15.4075C6.13274 15.554 6.05044 15.7527 6.05044 15.9598C6.05044 16.167 6.13274 16.3657 6.27922 16.5122L6.83154 15.9598ZM15.9595 6.83186L16.5118 6.27954C16.3654 6.13305 16.1667 6.05076 15.9595 6.05076C15.7524 6.05076 15.5537 6.13305 15.4072 6.27954L15.9595 6.83186ZM18.1688 9.04113L18.7211 9.59345C19.0261 9.28841 19.0261 8.79385 18.7211 8.48881L18.1688 9.04113ZM9.04081 18.1691L8.48849 18.7214C8.63498 18.8679 8.83365 18.9502 9.04081 18.9502C9.24797 18.9502 9.44665 18.8679 9.59313 18.7214L9.04081 18.1691ZM7.38386 16.5122L16.5118 7.38418L15.4072 6.27954L6.27922 15.4075L7.38386 16.5122ZM15.4072 7.38418L17.6165 9.59345L18.7211 8.48881L16.5118 6.27954L15.4072 7.38418ZM17.6165 8.48881L8.48849 17.6168L9.59313 18.7214L18.7211 9.59345L17.6165 8.48881ZM9.59313 17.6168L7.38386 15.4075L6.27922 16.5122L8.48849 18.7214L9.59313 17.6168Z" fill="currentColor"/>
+        </svg>
+    );
+}
+
+function TimeIcon() {
+    return (
+        <svg className="FooterIcon TimeIcon" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M1.96336 6.30713C1.74958 6.66192 1.86389 7.12283 2.21868 7.33661C2.57346 7.55038 3.03437 7.43607 3.24815 7.08129L1.96336 6.30713ZM2.5879 15.8197C2.43 15.4367 1.99156 15.2543 1.60863 15.4122C1.22569 15.5701 1.04326 16.0086 1.20116 16.3915L2.5879 15.8197ZM11.6058 2.35559C16.9905 2.35559 21.3558 6.72082 21.3558 12.1056L22.8558 12.1056C22.8558 5.89239 17.819 0.855592 11.6058 0.855592L11.6058 2.35559ZM21.3558 12.1056C21.3558 17.4904 16.9905 21.8556 11.6058 21.8556L11.6058 23.3556C17.819 23.3556 22.8558 18.3188 22.8558 12.1056L21.3558 12.1056ZM3.24815 7.08129C4.95544 4.24789 8.05998 2.35559 11.6058 2.35559L11.6058 0.855592C7.51256 0.855591 3.93077 3.04204 1.96336 6.30713L3.24815 7.08129ZM11.6058 21.8556C7.53732 21.8556 4.04917 19.3635 2.5879 15.8197L1.20116 16.3915C2.88598 20.4774 6.90887 23.3556 11.6058 23.3556L11.6058 21.8556Z" fill="currentColor"/>
+            <path d="M2.10477 4.10559L2.10477 7.10559L5.00342 6.62248" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+            <path d="M15 11.5L11 13.5V7" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+function FrontIcon() {
+    return (
+        <svg className="Icon FrontIcon" viewBox="0 0 512 512" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+            <path d="M 513 211.67 C 492.51 212 472.01 211.89 451.52 212.08 C 447.45 212.12 445.1 210.56 443.12 207.26 C 425.43 177.77 407.68 148.3 389.88 118.87 C 378.02 99.27 365.92 79.82 354.24 60.11 C 352.12 56.54 349.6 55.97 346.09 55.98 C 296.6 56.02 247.12 56 197.64 56 C 187.64 56 177.64 56.11 167.65 55.95 C 163.97 55.89 161.69 56.94 159.64 60.35 C 137.92 96.55 115.97 132.61 94.09 168.72 C 86.26 181.65 78.58 194.67 70.53 207.46 C 68.37 210.91 68.24 213.44 70.37 216.92 C 86.32 243 102.09 269.19 117.92 295.35 C 131.68 318.09 145.53 340.79 159.15 363.62 C 161.47 367.49 164.07 369.1 168.69 369.09 C 227.16 368.95 285.65 369.06 344.12 368.9 C 349.68 368.88 353.1 370.33 356.11 375.39 C 377.61 411.52 399.48 447.43 421.22 483.41 C 421.64 484.11 422 484.85 422.74 486.23 C 418.56 488.62 414.47 491.01 410.33 493.31 C 400.01 499.02 389.65 504.67 379.34 510.4 C 378.5 510.86 377.86 511.67 377.06 512.66 C 375.93 513 374.87 513 373.42 512.63 C 371.73 509.99 370.44 507.7 369.08 505.45 C 354.02 480.41 338.8 455.46 324.01 430.26 C 321.14 425.36 317.89 423.88 312.46 423.9 C 253.51 424.06 194.57 423.96 135.62 424.08 C 131.18 424.09 128.25 422.93 125.87 418.96 C 111.96 395.75 97.79 372.69 83.79 349.54 C 69.54 325.97 55.4 302.33 41.16 278.75 C 30.26 260.71 19.33 242.68 8.28 224.73 C 6.17 221.3 3.44 218.23 1 215 C 1 213.61 1 212.22 1.34 210.35 C 2.94 207.96 4.26 206.09 5.44 204.13 C 28.98 165.44 52.5 126.73 76.03 88.04 C 88.85 66.94 101.76 45.88 114.5 24.73 C 119.19 16.94 123.51 8.92 128 1 C 213.69 1 299.38 1 385.54 1.36 C 386.82 3.36 387.48 5.08 388.43 6.64 C 403.48 31.44 418.59 56.2 433.63 81.01 C 456.59 118.88 479.46 156.81 502.47 194.66 C 505.68 199.95 509.47 204.89 513 210 C 513 210.44 513 210.89 513 211.67 Z M 301.15 263.32 C 316.16 287.91 331.17 312.5 346.83 338.16 C 370.24 299.51 392.96 261.99 416.03 223.91 C 413.92 223.47 412.83 223.05 411.74 223.05 C 380.91 223.02 350.08 223.04 319.25 223 C 315.97 222.99 314.7 224.57 313.83 227.58 C 311.33 236.18 306.98 243.88 300.38 249.95 C 296.69 253.35 296.02 256.28 299.37 260.06 C 300.02 260.79 300.37 261.78 301.15 263.32 Z M 197.56 288.07 C 203.67 278.17 209.75 268.25 215.88 258.37 C 217.7 255.44 217.17 253.2 214.56 251.01 C 207.56 245.17 203.17 237.48 200.67 228.92 C 199.33 224.37 197.33 222.88 192.69 222.91 C 163.2 223.08 133.7 222.99 104.2 223.02 C 102.46 223.02 100.72 223.29 97.7 223.55 C 121.04 262.09 143.78 299.65 167.14 338.21 C 177.72 320.76 187.44 304.72 197.56 288.07 Z M 237.5 75 C 221.32 75 205.14 75 187.94 75 C 189.17 77.55 189.63 78.75 190.29 79.83 C 195.16 87.92 200.11 95.97 204.98 104.06 C 215.09 120.84 225.22 137.6 235.22 154.44 C 237.21 157.79 239.31 158.1 243.26 157.71 C 252.24 156.82 261.49 156.48 270.38 157.7 C 275.32 158.37 277.11 157.08 279.28 153.49 C 292.29 131.9 305.47 110.41 318.52 88.85 C 321.02 84.71 323.21 80.38 326.17 75 C 295.998 75 267.25 75 237.5 75 Z M 200.79 195.28 C 203.22 186.33 208.69 179.24 214.91 172.71 C 217.02 170.49 217.21 168.84 215.64 166.27 C 202.96 145.44 190.41 124.52 177.77 103.66 C 174.47 98.21 170.99 92.85 167 86.5 C 143.74 124.96 121.15 162.3 98.13 200.37 C 100.79 200.69 102.07 200.97 103.36 200.97 C 133.49 201 163.62 200.94 193.76 201.05 C 197.6 201.07 199.72 199.65 200.79 195.28 Z M 315.55 200.35 C 317.02 200.56 318.49 200.96 319.96 200.96 C 350.11 201 380.27 201 410.43 200.97 C 411.83 200.97 413.24 200.66 414.68 200.49 C 413.38 194.21 349.16 89.77 345.61 88.18 C 329.15 115.37 312.66 142.59 296.16 169.85 C 305.55 178.26 311.41 188.31 315.55 200.35 Z M 260.9 268.23 C 255 267.83 249.02 267.92 243.23 266.87 C 239.26 266.15 237.22 266.98 235.18 270.4 C 220.6 294.77 205.83 319.02 191.16 343.32 C 190.28 344.78 189.68 346.42 188.68 348.56 C 234.54 348.56 279.52 348.56 325 348.56 C 324.52 346.97 324.39 345.8 323.85 344.87 C 319.76 337.84 315.65 330.83 311.43 323.88 C 301.41 307.43 291.41 290.97 281.25 274.61 C 274.86 264.31 277.71 265.75 266.76 267.76 C 265.13 268.05 263.44 268 260.9 268.23 Z"/>
+        </svg>
+    );
+}
+
+const IconElements = {
+    HomeIcon,
+    RankingIcon,
+    ChallengeIcon,
+    GroupIcon,
+    UserIcon,
+    NotificationIcon,
+    SettingsIcon,
+    FitnessIcon,
+    TimeIcon,
+    FrontIcon
+};
+
+export default IconElements;
