+++ src/layoutElements.css
@@ -1,35 +1,35 @@
 /* 
-.hex:nth-child(even) => hex 클래스 중에서 짝수 순서인 것들만 선택
-:nth-child(odd) = 홀수번째
-:nth-child(3) = 정확히 3번째
-:nth-child(3n) = 3의 배수번째 (3, 6, 9...)
-:first-child = 첫번째
-:last-child = 마지막
-:hover = 마우스 올렸을 때
+.hex:nth-child(even) => Select only even-order hex classes
+:nth-child(odd) = Odd-order elements
+:nth-child(3) = Exactly the 3rd element
+:nth-child(3n) = Every 3rd element (3, 6, 9...)
+:first-child = First element
+:last-child = Last element
+:hover = When the mouse hovers over the element
 */
 
 /*
-부모 요소의 크기에 맞춰서 자식요소의 크기가 자동으로 조절되도록 만들고싶다면 Flexbox나 Grid를 사용할것.
-Flexbox의 경우 부모요소에서 display: flex를, 자식요소에서 flex:1 (자식 요소들을 동일한 비율로 분활) 사용.
-flex-direction을 같이 쓰면 배치 방향도 조절 가능
-aspect-ratio: 1/1 를 사용하면 각 그리드 아이템을 정사각형으로 고정 가능능
+If you want child elements to automatically adjust their size to match the parent element, use Flexbox or Grid.
+For Flexbox, use display: flex on the parent element and flex: 1 on the child elements (to divide them equally).
+You can also use flex-direction to adjust the layout direction.
+aspect-ratio: 1/1 can be used to fix each grid item as a square.
 */
 
 /*
-css에선 Dynamic Viewport Height, dvh라는게 있어서, 주소창이 보이거나 숨겨지는 것에 따라 실시간으로 변하는 뷰포트 높이를 설정가능.
-vh대신 쓰는 길이 단위.
-너비에는 dvw를 사용하면됨.
-svh 단위를 쓰면 주소창을 뺀 높이만 가져오는것도 가능함.
+CSS has Dynamic Viewport Height (dvh), which allows you to set viewport height that changes dynamically depending on whether the address bar is visible or hidden.
+Use dvh instead of vh for height.
+For width, use dvw.
+You can also use svh to get the height excluding the address bar.
 */
 
-/* 요소를 원형으로 배치하고싶으면 css transform이랑 삼각함수를 쓰는게 일반적. */
-/* 헥사곤은  clip-path: polygon(50% 0%, 100% 25%, 100% 75%, 50% 100%, 0% 75%, 0% 25%); 이걸로 만들수있음*/
+/* To arrange elements in a circular layout, use CSS transform and trigonometric functions. */
+/* Hexagon shapes can be created using clip-path: polygon(50% 0%, 100% 25%, 100% 75%, 50% 100%, 0% 75%, 0% 25%); */
 
 /*
-그림자
+Shadows
 box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
-            x y blur  색상(투명도포함)
-그림자를 넣고싶은 방향으로 x랑 y 오프셋을 px값으로 조절하면됨. 0이면 그 방향으로는 그림자 안생김.
+            x y blur color (including transparency)
+Adjust x and y offsets in px to control the direction of the shadow. If set to 0, no shadow appears in that direction.
 */
 
 :root {
@@ -37,7 +37,7 @@ box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
     --header-height: 56px;
     --footer-height: 80px; 
     --line-height: 2px;
-    --gaps: 4px; /* line 총합 */
+    --gaps: 4px; /* Total line spacing */
     --main-color: #96E36C;
     --light-color: #bcbcbcf3;
     --background-default: #3a3c32;
@@ -45,7 +45,7 @@ box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
     --background-red: #9f3434;
 }
 
-/* 모든 스크롤바 숨기기 */
+/* Hide all scrollbars */
 * {
     scrollbar-width: none;
     -ms-overflow-style: none;
@@ -64,13 +64,13 @@ body, html {
 }
 
 .MainContainer {
-    /* 컨텐츠 영역의 높이를 주소창이 있는 시점으로 고정시켜서, 주소창이 사라지면 여백을 늘리는걸로 컨텐츠 영역의 크기가 주소창 여부에 상관없이 유지되도록록 */
+    /* Fix the height of the content area based on the presence of the address bar, and adjust the padding to maintain the content area size regardless of the address bar visibility */
     display: grid;
     grid-template-rows: 
     var(--header-height)
     var(--line-height)
-    /* 중앙의 컨텐츠 영역의 크기를 계산 */
-    /* 주소창을 제외한 화면높이 - 헤더높이 - 푸터높이 - 여백 */
+    /* Calculate the size of the central content area */
+    /* Screen height excluding the address bar - header height - footer height - spacing */
     calc(100svh - var(--header-height) - var(--footer-height) - var(--gaps)) 
     var(--line-height)
     var(--footer-height);
